added userid in interviews support in C and R

+++ Routes/InterviewCalls.js
@@ -1,8 +1,27 @@
 const express = require('express');
 const router = express.Router();
 const jwt = require('jsonwebtoken');
+const User = require('../Models/User');
 const Interview = require('../Models/Interview');
 const tokenAuth = require('../Helpers/auth_mech');
+const userIdHelper = require('../Helpers/getuserid_mech');
+
+// ---------------------------- New User Sign Up
+router.post('/SignUp', function(req, res, next){ 
+    console.log("I am able to hit the SignUp API");
+    
+    User.create(req.body).then(function(signupresponse){
+        var userSignUpResponse = JSON.stringify({
+            "Message": "Your Signed-In Successfully",
+            "UserId": signupresponse._id
+        });
+
+        res.send(userSignUpResponse);
+
+    }).catch(next);
+    
+});
+
 
 // ---------------------------- Register a New Interview Call
 router.post('/RegisterNewInterview', tokenAuth.verifyToken, function(req, res, next){
@@ -13,21 +32,33 @@ router.post('/RegisterNewInterview', tokenAuth.verifyToken, function(req, res, n
         }
         else
         {
-            console.log("I have Entered Register New Interview API");
-            // console.log("\n"+"Request Body send to the API"+"\n");
-            // console.log(req.body);
-            Interview.create(req.body).then(function(interview){
-            
-            var registerInterviewResponse = JSON.stringify(
-                {
-                    "Interview Id": interview._id,
-                    "Company Name": interview.Company,
-                    "Interview Date": interview.Interview_Date,
-                    "Message": "Interview Details Saved Successfully"
-                });
-            
-            res.send(registerInterviewResponse);
-        }).catch(next);
+            userIdHelper.getUserIdFromToken(authData, function(userid){
+                
+                console.log("UserId in Interview js"+userid);
+    
+                    if(userid != null)
+                    {
+    
+                        console.log("I have Entered Register New Interview API");
+
+                        req.body.User_Id = userid.substring(1,userid.length-1);                        
+    
+                        Interview.create(req.body).then(function(interview){
+                        
+                        var registerInterviewResponse = JSON.stringify(
+                            {
+                                "Interview Id": interview._id,
+                                "Company Name": interview.Company,
+                                "Interview Date": interview.Interview_Date,
+                                "Message": "Interview Details Saved Successfully"
+                            });
+                        
+                        res.send(registerInterviewResponse);
+                        }).catch(next);
+                    } 
+            });
+
+
         }
     });
     
@@ -42,14 +73,18 @@ router.put('/UpdateInterviewDetails/:InterviewId', tokenAuth.verifyToken, functi
         }
         else
         {
+            userIdHelper.getUserIdFromToken(authData, function(userid){
+                
+            console.log("UserId in Interview js"+userid);
+
             console.log("I have Entered Update Interview Details API");
 
             var interviewId = req.params.InterviewId;
 
-            Interview.findByIdAndUpdate({_id: interviewId}, req.body).then(function(){
+            Interview.findByIdAndUpdate({User_Id: userid.substring(1,userid.length-1),_id: interviewId}, req.body).then(function(){
             
             // If we return the updatedInterview, then the Enteries would not be shown reflected, thus
-            Interview.findOne({_id: interviewId}).then(function(updatedInterview){
+            Interview.findOne({User_Id: userid.substring(1,userid.length-1),_id: interviewId}).then(function(updatedInterview){
 
                 var updateInterviewResponse = JSON.stringify(
                     {
@@ -61,8 +96,12 @@ router.put('/UpdateInterviewDetails/:InterviewId', tokenAuth.verifyToken, functi
                 res.send(updateInterviewResponse);
             }).catch(next);
             
-        }).catch(next);
+            }).catch(next);
+
+            });
+
         }
+    
     });
 
 });
@@ -76,13 +115,18 @@ router.delete('/RemoveSavedInterview/:InterviewId', tokenAuth.verifyToken, funct
         }
         else
         {
+            userIdHelper.getUserIdFromToken(authData, function(userid)
+            {
+                
+            console.log("UserId in Interview js"+userid);
+
             console.log("I have Entered Remove a Saved Interview API");
 
             // Grabbing the InterviewId for the Interview to be delete from the parameters in the URL
             var interviewId = req.params.InterviewId; 
         
             // Method will Find the Element with the Unique Object ID in Mongo DB and will remove it
-            Interview.findByIdAndRemove({_id: interviewId}).then(function(removedinterview){
+            Interview.findByIdAndRemove({User_Id: userid.substring(1,userid.length-1),_id: interviewId}).then(function(removedinterview){
 
             var deleteInterviewResponse = JSON.stringify(
                 {
@@ -92,6 +136,10 @@ router.delete('/RemoveSavedInterview/:InterviewId', tokenAuth.verifyToken, funct
             
             res.send(deleteInterviewResponse);
             }).catch(next);
+
+
+            });
+        
         }
     });
 
@@ -107,8 +155,17 @@ router.get('/ScheduledInterviews', tokenAuth.verifyToken, function(req, res, nex
             else
             {
                 console.log("I am Inside the Get All Scheduled Interviews API");
-    
-                Interview.find().sort({Interview_Date: -1}).then(function(interviews){
+                userIdHelper.getUserIdFromToken(authData, function(userid)
+                {
+                
+                console.log("UserId in Interview js"+userid);
+
+                if(userid != null)
+                {
+                    console.log({User_Id: userid.substring(1, userid.length-1)});
+                    
+                Interview.find({User_Id: userid.substring(1, userid.length-1)}).sort({Interview_Date: -1}).then(function(interviews){
+                    //.sort({Interview_Date: -1})
                     
                     // Returning Interviews Sorted in Descending Order of the Interview Date
                     res.send(
@@ -118,6 +175,10 @@ router.get('/ScheduledInterviews', tokenAuth.verifyToken, function(req, res, nex
                         }); 
         
                 }).catch(next);
+                
+                }
+                });
+
             }
         });
 
@@ -129,21 +190,28 @@ router.get('/ScheduledInterviews', tokenAuth.verifyToken, function(req, res, nex
 router.get('/ScheduledInterviews/:InterviewDate', tokenAuth.verifyToken, function(req, res, next){
         
         jwt.verify(req.token, 'secretkey', (err, authData)=>{
-            if(err)
-            {
+            
+        if(err)
+        {
             res.sendStatus(403);
-            }
-            else
-            {
+        }
+        else
+        {
+            userIdHelper.getUserIdFromToken(authData, function(userid){
+                
+            console.log("UserId in Interview js"+userid);
+
             console.log("I am Inside the Get All Scheduled Interviews By Date API");
 
             var interviewDate = req.params.InterviewDate;
 
-            Interview.find({Interview_Date: interviewDate}).then(function(interviewsondate){
+            Interview.find({User_Id: userid.substring(1,userid.length-1),Interview_Date: interviewDate}).then(function(interviewsondate){
             
             res.send(interviewsondate); 
 
             }).catch(next);
+
+            });
         }
     });
 
@@ -160,16 +228,23 @@ router.get('/GetInterview/:InterviewId', tokenAuth.verifyToken, function(req, re
         }
         else
         {
-        console.log("I am Inside the Get Interviews By Id API");
+            userIdHelper.getUserIdFromToken(authData, function(userid){
+                
+            console.log("UserId in Interview js"+userid);
 
-        //var id = req.params.InterviewId;
+            console.log("I am Inside the Get Interviews By Id API");
 
-        Interview.findById(req.params.InterviewId).then(function(interview){
+            //var id = req.params.InterviewId;
+
+            Interview.find({User_Id: userid.substring(1,userid.length-1)}).findById(req.params.InterviewId).then(function(interview){
         
-        res.send(interview); 
+            res.send(interview); 
 
-        }).catch(next);
-    }
+            }).catch(next);
+
+            });
+
+        }
 });
 
 });
